requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
google-generativeai==0.6.0
